Use int not long

+++ src/Twilio.TaskRouter/Model/TaskQueueStatistics.cs
@@ -97,7 +97,7 @@ namespace Twilio.TaskRouter
             /// <summary>
             /// Gets or sets the tasks by priority.
             /// </summary>
-            public Dictionary<string, long> TasksByPriority { get; set; }
+            public Dictionary<string, int> TasksByPriority { get; set; }
             /// <summary>
             /// Gets or sets the activity statistics.
             /// </summary>
